restructured databse files and model
