+++ apps/backend/src/database/models.ts
@@ -1,0 +1,5 @@
+import {UserSchema, IUser} from "./user/user.schema"
+
+import {model} from "mongoose"
+
+export const User = model<IUser>('User', UserSchema);
