--- apps/backend/src/database/models/models.ts
@@ -1,4 +1,0 @@
-import UserSchema from "./user/user.schema"
-import {model} from "mongoose"
-
-export const User = model('User', UserSchema);
