--- apps/backend/src/database/models/user/user.schema.ts
@@ -1,8 +1,0 @@
-import { Schema } from "mongoose";
-
-const UserSchema = new Schema({
-	username: String,
-	hashed_password: String
-});
-
-export default UserSchema;
