+++ apps/backend/src/database/user/user.schema.ts
@@ -1,0 +1,12 @@
+import { Schema, Document } from "mongoose";
+
+export interface IUser extends Document {
+	username: string;
+	hashed_password: string;
+}
+
+
+export const UserSchema: Schema = new Schema({
+	username: String,
+	hashed_password: String
+});
